feat: package media manager as an executable
