+++ package.json
@@ -28,10 +28,17 @@
   ],
   "scripts": {
     "info": "npm-scripts-info",
-    "build": "trash dist && yarn build:main",
+    "build": "trash dist && yarn build:main && yarn generate-deps-metadata",
     "buildstart": "yarn build && yarn start",
     "buildinspect": "yarn build && yarn inspect",
     "build:main": "tsc -p tsconfig.json",
+    "generate-deps-metadata": "node generate-deps-metadata.js",
+    "remove-fauxton-package": "json -I -f node_modules/express-pouchdb/package.json -e \"this.dependencies[\\\"pouchdb-fauxton\\\"] = undefined\"",
+    "remove-fauxton-win32": "npm run remove-fauxton-package && del node_modules\\express-pouchdb\\lib\\routes\\fauxton.js",
+    "remove-fauxton-linux": "npm run remove-fauxton-package && (rm node_modules/express-pouchdb/lib/routes/fauxton.js || true)",
+    "build-linux": "(rm -rf dist || true) && yarn build && npm run remove-fauxton-linux && pkg -t node10-linux dist/index.js -o dist/media-manager && cp ./node_modules/leveldown/build/Release/leveldown.node ./dist",
+    "build-win32-on-linux": "(rm -rf dist || true) && yarn build && pkg -t node10-win dist/index.js -o dist/media-manager.exe && cp ./node_modules/leveldown/build/Release/leveldown.node ./dist",
+    "build-win32": "rmdir /s /q dist & mkdir dist && yarn build && pkg -t node10-win dist/index.js -o dist/media-manager.exe && copy .\\node_modules\\leveldown\\build\\Release\\leveldown.node .\\dist /Y",
     "lint": "tslint --project tsconfig.jest.json --config tslint.json",
     "unit": "jest",
     "test": "yarn lint && yarn unit",
@@ -86,11 +93,13 @@
     "cpx": "^1.5.0",
     "gh-pages": "^1.2.0",
     "jest": "^23.1.0",
+    "json": "^9.0.6",
     "mkdirp": "^0.5.1",
     "node-license-validator": "^1.3.0",
     "npm-scripts-info": "^0.3.7",
     "nyc": "^12.0.2",
     "opn-cli": "^3.1.0",
+    "pkg": "^4.3.7",
     "pouchdb-adapter-memory": "^7.0.0",
     "sleep-ms": "^2.0.1",
     "standard-version": "^4.4.0",
